backend/pyproject.toml
[project]
name = "campus-admin-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Saad Sultan",email = "muhammadsaadsultan99@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.117.1,<0.118.0)",
    "uvicorn[standard] (>=0.36.0,<0.37.0)",
    "motor (>=3.7.1,<4.0.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "openai (>=1.108.1,<2.0.0)",
    "langgraph (>=0.6.7,<0.7.0)",
    "langchain-core (>=0.3.76,<0.4.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "websockets (>=15.0.1,<16.0.0)",
    "langgraph-store-mongodb (>=0.1.0,<0.2.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
